Move shortcut.ts
