+++ packages/client/src/component/shortcut.ts
